+++ python/GafferTest/FileSystemPathTest.py
@@ -39,7 +39,7 @@ import unittest
 import time
 import datetime
 import os
-if os.name is not "nt":
+if os.name is not "nt" :
 	import pwd
 	import grp
 
@@ -124,11 +124,13 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 
 		self.assertEqual( p, p2 )
 		self.assertEqual( str( p ), str( p2 ) )
+		self.assertEqual( p.nativeString(), p2.nativeString() )
 
 	def testEmptyPath( self ) :
 
 		p = Gaffer.FileSystemPath()
 		self.assertEqual( str( p ), "" )
+		self.assertEqual( p.nativeString(), "" )
 		self.assertTrue( p.isEmpty() )
 		self.assertFalse( p.isValid() )
 
@@ -136,26 +138,79 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 
 		os.chdir( self.temporaryDirectory() )
 
-		with open( self.temporaryDirectory() + "/a", "w" ) as f :
+		with open( os.path.join( self.temporaryDirectory(), "a" ), "w" ) as f :
 			f.write( "AAAA" )
 
 		p = Gaffer.FileSystemPath( "a" )
 
 		self.assertEqual( str( p ), "a" )
+		self.assertEqual( p.nativeString(), "a" )
 		self.assertFalse( p.isEmpty() )
 		self.assertTrue( p.isValid() )
 
 		p2 = Gaffer.FileSystemPath( "nonexistent" )
 
 		self.assertEqual( str( p2 ), "nonexistent" )
+		self.assertEqual( p2.nativeString(), "nonexistent" )
 		self.assertFalse( p2.isEmpty() )
 		self.assertFalse( p2.isValid() )
+
+	def testDriveLetterPath( self ) :
+		p = Gaffer.FileSystemPath( "C:\\this\\path\\does\\not\\exist.ext" )
+
+		self.assertEqual( str( p ), "C:/this/path/does/not/exist.ext" )
+		if os.name == "nt" :
+			self.assertEqual( p.nativeString(), "C:\\this\\path\\does\\not\\exist.ext")
+		else :
+			self.assertEqual( p.nativeString(), str( p ) )
+		self.assertFalse( p.isEmpty() )
+		self.assertFalse( p.isValid() )
+
+		p = Gaffer.FileSystemPath( "/C:/leading/slash/should/be/removed" )
+		self.assertEqual( str( p ), "C:/leading/slash/should/be/removed" )
+
+	def testUNCPath( self ) :
+		p = Gaffer.FileSystemPath( "\\\\this.server\\path\\does\\not\\exist.ext" )
+
+		self.assertEqual( str( p ), "//this.server/path/does/not/exist.ext" )
+		if os.name == "nt" :
+			self.assertEqual( p.nativeString(), "\\\\this.server\\path\\does\\not\\exist.ext")
+		else :
+			self.assertEqual( p.nativeString(), str( p ) )
+		self.assertFalse( p.isEmpty() )
+		self.assertFalse( p.isValid() )
+
+	def testPosixPath( self ) :
+		p = Gaffer.FileSystemPath( "/this/path/does/not/exist.ext" )
+
+		if os.name == "nt" :
+			self.assertEqual( str( p ), "//this/path/does/not/exist.ext" )
+			self.assertEqual( p.nativeString(), "\\\\this\\path\\does\\not\\exist.ext" )
+		else :
+			self.assertEqual( str( p ), "/this/path/does/not/exist.ext" )
+			self.assertEqual( p.nativeString(), "/this/path/does/not/exist.ext")
+
+		self.assertFalse( p.isEmpty() )
+		self.assertFalse( p.isValid() )
+
+	def testMixedSlashes( self ) :
+		p = Gaffer.FileSystemPath( "oh\\my/what/a\\mess" )
+		self.assertEqual( str( p ), "oh/my/what/a/mess" )
+		
+		p = Gaffer.FileSystemPath( "C:\\not/a\\drive/letter\\too" )
+		self.assertEqual( str( p ), "C:/not/a/drive/letter/too" )
+
+		p = Gaffer.FileSystemPath( "/unc\\paths/must\\also/work" )
+		if os.name == "nt" :
+			self.assertEqual( str( p ), "//unc/paths/must/also/work" )
+		else :
+			self.assertEqual( str( p ), "/unc/paths/must/also/work" )
 
 	def testRelativePathChildren( self ) :
 
 		os.chdir( self.temporaryDirectory() )
 		os.mkdir( "dir" )
-		with open( self.temporaryDirectory() + "/dir/a", "w" ) as f :
+		with open( os.path.join( self.temporaryDirectory(), "dir", "a" ), "w" ) as f :
 			f.write( "AAAA" )
 
 		p = Gaffer.FileSystemPath( "dir" )
@@ -163,6 +218,11 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 		c = p.children()
 		self.assertEqual( len( c ), 1 )
 		self.assertEqual( str( c[0] ), "dir/a" )
+		if os.name == "nt" :
+			self.assertEqual( c[0].nativeString(), "dir\\a" )
+		else:
+			self.assertEqual( c[0].nativeString(), "dir/a" )
+
 		self.assertTrue( c[0].isValid() )
 
 	def testChildrenOfFile( self ) :
@@ -175,7 +235,7 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 		p = Gaffer.FileSystemPath( self.temporaryDirectory() )
 		p.append( "t" )
 
-		with open( str( p ), "w" ) as f :
+		with open( p.nativeString(), "w" ) as f :
 			f.write( "AAAA" )
 
 		mt = p.property( "fileSystem:modificationTime" )
@@ -184,7 +244,7 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 
 		time.sleep( 1 )
 
-		with open( str( p ), "w" ) as f :
+		with open( p.nativeString(), "w" ) as f :
 			f.write( "BBBB" )
 
 		mt = p.property( "fileSystem:modificationTime" )
@@ -196,19 +256,20 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 		p = Gaffer.FileSystemPath( self.temporaryDirectory() )
 		p.append( "t" )
 
-		with open( str( p ), "w" ) as f :
+		with open( p.nativeString(), "w" ) as f :
 			f.write( "AAAA" )
 
 		o = p.property( "fileSystem:owner" )
 		self.assertTrue( isinstance( o, str ) )
-		self.assertEqual( o, pwd.getpwuid( os.stat( str( p ) ).st_uid ).pw_name )
+
+		self.assertEqual( o, self.getFileOwner( p.nativeString() ) )
 
 	def testGroup( self ) :
 
 		p = Gaffer.FileSystemPath( self.temporaryDirectory() )
 		p.append( "t" )
 
-		with open( str( p ), "w" ) as f :
+		with open( p.nativeString(), "w" ) as f :
 			f.write( "AAAA" )
 
 		g = p.property( "fileSystem:group" )
@@ -216,7 +277,7 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 		if os.name != "nt" :
 			self.assertEqual( g, grp.getgrgid( os.stat( p.nativeString() ).st_gid ).gr_name )
 		else :
-			self.assertEQual( g, "" )
+			self.assertEqual( g, "" )
 
 	def testPropertyNames( self ) :
 
@@ -236,9 +297,9 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 
 	def testSequences( self ) :
 
-		os.mkdir( self.temporaryDirectory() + "/dir" )
+		os.mkdir( os.path.join( self.temporaryDirectory(), "dir" ) )
 		for n in [ "singleFile.txt", "a.001.txt", "a.002.txt", "a.004.txt", "b.003.txt" ] :
-			with open( self.temporaryDirectory() + "/" + n, "w" ) as f :
+			with open( os.path.join( self.temporaryDirectory(), n ), "w" ) as f :
 				f.write( "AAAA" )
 
 		p = Gaffer.FileSystemPath( self.temporaryDirectory(), includeSequences = True )
@@ -248,30 +309,38 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 		self.assertEqual( len( c ), 8 )
 
 		s = sorted( c, key=str )
-		self.assertEqual( str(s[0]), self.temporaryDirectory() + "/a.###.txt" )
-		self.assertEqual( str(s[1]), self.temporaryDirectory() + "/a.001.txt" )
-		self.assertEqual( str(s[2]), self.temporaryDirectory() + "/a.002.txt" )
-		self.assertEqual( str(s[3]), self.temporaryDirectory() + "/a.004.txt" )
-		self.assertEqual( str(s[4]), self.temporaryDirectory() + "/b.###.txt" )
-		self.assertEqual( str(s[5]), self.temporaryDirectory() + "/b.003.txt" )
-		self.assertEqual( str(s[6]), self.temporaryDirectory() + "/dir" )
-		self.assertEqual( str(s[7]), self.temporaryDirectory() + "/singleFile.txt" )
+		self.assertEqual( str(s[0]), self.temporaryDirectory().replace( "\\", "/" ) + "/a.###.txt" )
+		self.assertEqual( s[0].nativeString(), os.path.join( self.temporaryDirectory(), "a.###.txt" ) )
+		self.assertEqual( str(s[1]), self.temporaryDirectory().replace( "\\", "/" ) + "/a.001.txt" )
+		self.assertEqual( s[1].nativeString(), os.path.join( self.temporaryDirectory(), "a.001.txt" ) )
+		self.assertEqual( str(s[2]), self.temporaryDirectory().replace( "\\", "/" ) + "/a.002.txt" )
+		self.assertEqual( s[2].nativeString(), os.path.join( self.temporaryDirectory(), "a.002.txt" ) )
+		self.assertEqual( str(s[3]), self.temporaryDirectory().replace( "\\", "/" ) + "/a.004.txt" )
+		self.assertEqual( s[3].nativeString(), os.path.join( self.temporaryDirectory(), "a.004.txt" ) )
+		self.assertEqual( str(s[4]), self.temporaryDirectory().replace( "\\", "/" ) + "/b.###.txt" )
+		self.assertEqual( s[4].nativeString(), os.path.join( self.temporaryDirectory(), "b.###.txt" ) )
+		self.assertEqual( str(s[5]), self.temporaryDirectory().replace( "\\", "/" ) + "/b.003.txt" )
+		self.assertEqual( s[5].nativeString(), os.path.join( self.temporaryDirectory(), "b.003.txt" ) )
+		self.assertEqual( str(s[6]), self.temporaryDirectory().replace( "\\", "/" ) + "/dir" )
+		self.assertEqual( s[6].nativeString(), os.path.join( self.temporaryDirectory(), "dir" ) )
+		self.assertEqual( str(s[7]), self.temporaryDirectory().replace( "\\", "/" ) + "/singleFile.txt" )
+		self.assertEqual( s[7].nativeString(), os.path.join( self.temporaryDirectory(), "singleFile.txt" ) )
 
 		for x in s :
 
 			self.assertTrue( x.isValid() )
-			if not os.path.isdir( str(x) ) :
+			if not os.path.isdir( x.nativeString() ) :
 				self.assertTrue( x.isLeaf() )
 
 			self.assertEqual( x.property( "fileSystem:owner" ), self.getFileOwner( p.nativeString() ) )
-			if os.name is not "nt":
+			if os.name is not "nt" :
 				self.assertEqual( x.property( "fileSystem:group" ), grp.getgrgid( os.stat( str( p ) ).st_gid ).gr_name )
 			self.assertLess( (datetime.datetime.utcnow() - x.property( "fileSystem:modificationTime" )).total_seconds(), 2 )
-			if "###" not in str(x) :
+			if "###" not in str( x ) :
 				self.assertFalse( x.isFileSequence() )
 				self.assertEqual( x.fileSequence(), None )
 				self.assertEqual( x.property( "fileSystem:frameRange" ), "" )
-				if os.path.isdir( str(x) ) :
+				if os.path.isdir( x.nativeString() ) :
 					self.assertEqual( x.property( "fileSystem:size" ), 0 )
 				else :
 					self.assertEqual( x.property( "fileSystem:size" ), 4 )
@@ -279,13 +348,13 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 		self.assertEqual( s[0].property( "fileSystem:frameRange" ), "1-2,4" )
 		self.assertTrue( s[0].isFileSequence() )
 		self.assertTrue( isinstance( s[0].fileSequence(), IECore.FileSequence ) )
-		self.assertEqual( s[0].fileSequence(), IECore.FileSequence( str(s[0]), IECore.frameListFromList( [ 1, 2, 4 ] ) ) )
+		self.assertEqual( s[0].fileSequence(), IECore.FileSequence( s[0].nativeString(), IECore.frameListFromList( [ 1, 2, 4 ] ) ) )
 		self.assertEqual( s[0].property( "fileSystem:size" ), 4 * 3 )
 
 		self.assertEqual( s[4].property( "fileSystem:frameRange" ), "3" )
 		self.assertTrue( s[4].isFileSequence() )
 		self.assertTrue( isinstance( s[4].fileSequence(), IECore.FileSequence ) )
-		self.assertEqual( s[4].fileSequence(), IECore.FileSequence( str(s[4]), IECore.frameListFromList( [ 3 ] ) ) )
+		self.assertEqual( s[4].fileSequence(), IECore.FileSequence( s[4].nativeString(), IECore.frameListFromList( [ 3 ] ) ) )
 		self.assertEqual( s[4].property( "fileSystem:size" ), 4 )
 
 		# make sure we can copy
@@ -301,12 +370,18 @@ class FileSystemPathTest( GafferTest.TestCase ) :
 		self.assertEqual( len( c ), 6 )
 
 		s = sorted( c, key=str )
-		self.assertEqual( str(s[0]), self.temporaryDirectory() + "/a.001.txt" )
-		self.assertEqual( str(s[1]), self.temporaryDirectory() + "/a.002.txt" )
-		self.assertEqual( str(s[2]), self.temporaryDirectory() + "/a.004.txt" )
-		self.assertEqual( str(s[3]), self.temporaryDirectory() + "/b.003.txt" )
-		self.assertEqual( str(s[4]), self.temporaryDirectory() + "/dir" )
-		self.assertEqual( str(s[5]), self.temporaryDirectory() + "/singleFile.txt" )
+		self.assertEqual( str(s[0]), self.temporaryDirectory().replace( "\\", "/" ) + "/a.001.txt" )
+		self.assertEqual( s[0].nativeString(), os.path.join( self.temporaryDirectory(), "a.001.txt" ) )
+		self.assertEqual( str(s[1]), self.temporaryDirectory().replace( "\\", "/" ) + "/a.002.txt" )
+		self.assertEqual( s[1].nativeString(), os.path.join( self.temporaryDirectory(), "a.002.txt" ) )
+		self.assertEqual( str(s[2]), self.temporaryDirectory().replace( "\\", "/" ) + "/a.004.txt" )
+		self.assertEqual( s[2].nativeString(), os.path.join( self.temporaryDirectory(), "a.004.txt" ) )
+		self.assertEqual( str(s[3]), self.temporaryDirectory().replace( "\\", "/" ) + "/b.003.txt" )
+		self.assertEqual( s[3].nativeString(), os.path.join( self.temporaryDirectory(), "b.003.txt" ) )
+		self.assertEqual( str(s[4]), self.temporaryDirectory().replace( "\\", "/" ) + "/dir" )
+		self.assertEqual( s[4].nativeString(), os.path.join( self.temporaryDirectory(), "dir" ) )
+		self.assertEqual( str(s[5]), self.temporaryDirectory().replace( "\\", "/" ) + "/singleFile.txt" )
+		self.assertEqual( s[5].nativeString(), os.path.join( self.temporaryDirectory(), "singleFile.txt" ) )
 
 		# and we can include them again
 		p.setIncludeSequences( True )